requirements.txt
# FastAPI + Server
fastapi==0.110.0
uvicorn==0.29.0

# LangChain ecosystem
langchain==0.1.13
langchain-core==0.1.32
langchain-community==0.0.28
langchain-openai==0.1.2

# LLM client (OpenRouter / OpenAI compatible)
openai==1.14.2

# NLP + RAG stack
sentence-transformers==2.2.2
transformers==4.38.2
torch==2.1.2
faiss-cpu==1.7.4

# PDF processing
pdfplumber==0.11.0
pdfminer-six==20231228

# Web scraping
beautifulsoup4==4.12.3
requests==2.31.0

# Database
pymysql==1.1.0
SQLAlchemy==2.0.29

# Utilities
python-dotenv==1.0.1
tqdm==4.66.2
numpy==1.26.4
pandas==2.2.1
regex==2023.12.25
